pyproject.toml
[project]
name = "sanaap-backend-challenge-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.8",
    "django-filter>=25.2",
    "django-storages[s3]>=1.14.6",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "uvicorn[standard]>=0.38.0",
]
